refactor: :recycle: extract function to esmodule
